Majorly changed the way that the library is used - for added simplicity
A new Processable class has been added which is to be extended, rather than the WebSocket class to avoid calling the parent's constructor.
A pointer to this new Processable class is to be passed to WebSocket::start( ) if no paramaters are passed then the default Processable object will be passed.
You no longer need to supply the WebSocket class with a boost::asio::io_stream object as it now uses its own, which also removes the need to cal io_stream::run( )

See chatbot.cpp for example